refactor test module
