+++ pom.xml
@@ -17,8 +17,9 @@
     </modules>
 
     <properties>
+        <project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
         <lombok.version>1.18.8</lombok.version>
-    </properties>
+</properties>
 
     <build>
         <plugins>
