+++ test/src/test/java/test/org/dst/core/operator/KVSTableTest.java
@@ -1,0 +1,13 @@
+package test.org.dst.core.operator;
+import org.junit.jupiter.api.Test;
+
+
+
+public class KVSTableTest {
+
+
+    @Test
+    public void testTable() {
+    }
+
+}
